Update cidr.go
Currently, the "-ip-format" flag for case 5 does not properly work. For example, as reflected in ticket 221, the command `echo '127.0.0.1' | mapcidr -ip-format 5` incorrectly returns the value `281472812449793` when it should return the value 
`2130706433`

This pull request fixes the function "IPToInteger" to correctly convert both IPv4 and IPv6 addresses to their respective integer representations.